separacion de responsabilidades
